removed picamera2 interval focuser. support only continuous mode from now on.
